requirement.txt
torch
torchvision
numpy
sacred
Pillow
visdom
csv
kornia